Update Mbed OS integration
This commit updates the way this driver is integrated in Mbed OS.
It adds a Python script, mbed-os.py that deletes some files to use the
driver in 800-90B TRNG mode. It also updates the README.md with the new
instructions.
The pull request ARMmbed/mbed-os#6169 was merged in Mbed OS using this
new way.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>